\name{ionScore-methods}
\docType{methods}
\alias{ionScore}
\alias{ionScore-methods}

\title{ Accessor methods of the ion score in the \code{scores} slot }

\description{
  This method allows to specifically retrieve the ion score of an
  object of class \code{"\linkS4class{SearchHit}"}. All scores can be
  retrieved with the \code{scores} method (see
  \code{\link{scores-methods}}) as a named numeric.
}
\section{Methods}{
  \describe{
    \item{\code{signature(object = "SearchHit")}}{
      object of class \code{"\linkS4class{SpectrumQuery}"}
    }
  }
}

\keyword{methods}
