\name{modifications-methods}
\docType{methods}
\alias{modifications-methods}
\alias{modifications<-}
\alias{modifications}

\title{ Accessor methods of the \code{modifications} slot }

\description{
  These methods allow to retrieve or update the \code{modification} slot
  of an instance of \code{"\linkS4class{SearchHit}"}. The values are
  set automatically when creating an instance of
  \code{"\linkS4class{MSMSpepXML}"} with \code{\link{parseMSMSpepXML}},
  and should generally not need to be changed.
}

\section{Methods}{
  \describe{
    \item{\code{signature(object = "SearchHit")}}{
      object of class \code{"\linkS4class{SearchHit}"}
    }
  }
}

\keyword{methods}

