\name{nSpectrumQueries-methods}
\docType{methods}
\alias{nSpectrumQueries}
\alias{nSpectrumQueries-methods}

\title{ Returns the numbre of search queries. }

\description{
  This methods returns the number of spectrum queries in an
  \code{"\linkS4class{MSMSpepXML}"} object.
}
\section{Methods}{
  \describe{
    \item{\code{signature(object = "MSMSpepXML")}}{
      an object of class \code{"\linkS4class{MSMSpepXML}"} is used as
      parameter and an integer is returned.
    }
  }
}

\keyword{methods}

